+++ spec/factories/tasks.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :task do
-    name { Admin::TasksController::TASKS_SEQUENCE.sample }
+    name { ClaimCheckingTasks.new(claim).applicable_task_names.sample }
     association :created_by, factory: :dfe_signin_user
     association :claim
   end