+++ spec/features/admin_checks_maths_and_physics_claim_spec.rb
@@ -15,6 +15,7 @@ RSpec.feature "Admin checking a Maths & Physics claim" do
 
     expect(page).to have_content("1. Qualifications")
     expect(page).to have_content("2. Employment")
+    expect(page).to have_content("3. Decision")
 
     click_on "Check qualification information"
     expect(page).to have_content("Qualifications")